Issue number:  113 Obtained from: kcr Submitted by:  kcr Reviewed by:   paulby
Temporary workaround for Issue 113 (disable transparency for GeometryArray
with per-vertex color on 2nd screen) until a real fix can be found.

Fixed a typo in the setTextureCoordinates(...,TexCoord3f[]) method.
It checked and threw an exception if the geometry array had 3D texcoords,
when it should have been checking for 4D texcoords.