Issue number:  144
Fixed Issue 144 : ExponentialFog blending factor depends on canvas size.

NOTE: As a result of this fix, applications will now need to recalculate
their exponential fog values for all Java 3D scenes. This is needed
because we are now correctly transforming the density into Eye Coordinates
(EC) where we were not doing so before. On a a 1280x1024 screen, density
values should now be set to approximately 1/5 of what they were in
previous versions of Java 3D.