Issue number:  147 Obtained from: robertf Submitted by:  kcr Reviewed by:   kcr
Ant file changes for issue 147: Need Windows/amd64 port of Java 3D

NOTE: we do not yet have a Windows64 machine setup to do builds or testing. Also, it is likely that more changes could be needed in the DirectX pipeline since that code was never verified to be 64-bit clean.

Also note: we still need to change the native pipeline "window" parameter from a 32-bit int to a 64-bit long. I will do that in a separate putback.